Evidencia_03

+++ backend/app/Http/Controllers/BootcampController.php
@@ -7,9 +7,10 @@ use App\Models\Bootcamp;
 use App\Http\Requests\StoreBootcampRequest;
 use App\Http\Resources\BootcampResource;
 use App\Http\Resources\BootcampCollection;
+use App\Http\Controllers\BaseController;
 
 
-class BootcampController extends Controller
+class BootcampController extends BaseController
 {
     /**
      * Display a listing of the resource.
@@ -21,8 +22,16 @@ class BootcampController extends Controller
         //json es un obejto que trae los metodos response
         //parametros: 1. data a enviar al client
                     //2. codigo status http
+         try{
+            return $this->sendResponse(new BootcampCollection(Bootcamp::all()));
+         }catch(\Exception $e){
+            return $this->sendError('Serve Error' ,500);
+         } 
+
                     
-     return response()->json( new BootcampCollection(Bootcamp::all()) ,200);
+
+
+    /* return response()->json( new BootcampCollection(Bootcamp::all()) ,200);*/
     }
 
     /**
@@ -33,13 +42,19 @@ class BootcampController extends Controller
      */
     public function store(StoreBootcampRequest $request)
     {
+        try {
+            return $this->sendResponse(new BootcampResource(Bootcamp::create($request->all())) , 201);
+    }
+        catch (\Exception $th) {
+            return $this->sendError('Server Error' , 500);
+        }
        //1. reglas de validacion
        
        //2. Crear el validador
        
        //3. Validar 
       
-        return response()->json([ "success" => true, "data" => new BootcampResource(Bootcamp::create($request->all()))] , 201);
+       /* return response()->json([ "success" => true, "data" => new BootcampResource(Bootcamp::create($request->all()))] , 201);*/
     }
 
     /**
@@ -50,7 +65,22 @@ class BootcampController extends Controller
      */
     public function show($id)
     {
-        return response()->json([ "success" => true, "data" => new BootcampResource(Bootcamp::find($id))] ,200);
+
+        try{
+           //1. Encontrar el bootcamp por id
+        $bootcamp = Bootcamp::find($id);
+        //2. en caso de que el bootcamp
+        // no exista
+        if(!$bootcamp){
+            return $this->sendError("bootcamp with id:$id not found", 400);
+        }
+        return $this->sendResponse(new BootcampResource($bootcamp)); 
+        }catch(\Exception $e){
+            return $this->sendError('Serve Error' ,500) ;
+        }
+
+
+        /*return response()->json([ "success" => true, "data" => new BootcampResource(Bootcamp::find($id))] ,200);*/
     }
 
     /**
@@ -62,11 +92,32 @@ class BootcampController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //Localizar el bootcamp con id
+        try {
+            //Localizar el bootcamp con id
         $b = bootcamp::find($id);
+        if(!$b){
+            return $this->sendError("bootcamp with id:$id not found", 400);
+        }
         //actualizarlo con update
         $b->update($request->all());
-        return response()->json(["success" => true, "data" => new BootcampResource($b)], 200);
+        return $this->sendResponse(new BootcampResource($b));
+        } catch (\Exception $th) {
+            return $this->sendError('Serve Error' ,500);
+        }
+
+        /*
+        //Localizar el bootcamp con id
+        $b = bootcamp::find($id);
+        if(!$b){
+            return $this->sendError("bootcamp with id:$id not found", 400);
+        }
+        //actualizarlo con update
+        $b->update($request->all());
+        return $this->sendResponse(new BootcampResource($b));*/
+        
+        //actualizarlo con update
+        /*$b->update($request->all());
+        return response()->json(["success" => true, "data" => new BootcampResource($b)], 200);*/
     }
 
     /**
@@ -77,10 +128,28 @@ class BootcampController extends Controller
      */
     public function destroy($id)
     {
-   
+        try {
+            $a = bootcamp::find($id);
+         if(!$a){
+            return $this->sendError("bootcamp with id:$id not found", 400);
+        }
+        //eliminar
+        $a->delete();
+        return $this->sendResponse(new BootcampResource($a));
+        } catch (\Throwable $th) {
+            return $this->sendError('Serve Error' ,500);
+        }
+
+        /* 
          $a = bootcamp::find($id);
-   
-         $a->delete();
-         return response()->json(["success" => true, "data" => new BootcampResource($a)], 200);
+         if(!$a){
+            return $this->sendError("bootcamp with id:$id not found", 400);
+        }
+        //eliminar
+        $a->delete();
+        return $this->sendResponse(new BootcampResource($a));*/
+
+        /* $a->delete();
+         return response()->json(["success" => true, "data" => new BootcampResource($a)], 200);*/
     }
 }
